Update TP Data Member Format.

+++ python/trgtools/TPReader.py
@@ -26,14 +26,14 @@ class TPReader(HDF5Reader):
     # TP data type
     tp_dt = np.dtype([
                       ('adc_integral', np.uint32),
-                      ('adc_peak', np.uint32),
-                      ('channel', np.int32),
-                      ('detid', np.uint16),
-                      ('flag', np.uint16),
-                      ('time_over_threshold', np.uint64),
-                      ('time_peak', np.uint64),
+                      ('adc_peak', np.uint16),
+                      ('channel', np.uint32),
+                      ('detid', np.uint8),
+                      ('flag', np.uint8),
+                      ('time_over_threshold', np.uint16),
+                      ('time_peak', np.uint16),
                       ('time_start', np.uint64),
-                      ('version', np.uint16)
+                      ('version', np.uint8)
                      ])
 
     def __init__(self, filename: str, verbosity: int = 0) -> None:
